+++ src/main/java/com/arangodb/internal/ArangoDBImpl.java
@@ -114,9 +114,12 @@ public class ArangoDBImpl extends InternalArangoDB<ArangoExecutorSync> implement
     public void shutdown() throws ArangoDBException {
         try {
             executor.disconnect();
-            cp.close();
-        } catch (final IOException e) {
-            throw new ArangoDBException(e);
+        } finally {
+            try {
+                cp.close();
+            } catch (final IOException e) {
+                LOGGER.error("Got exception during shutdown:", e);
+            }
         }
     }
 
@@ -169,6 +172,11 @@ public class ArangoDBImpl extends InternalArangoDB<ArangoExecutorSync> implement
     @Override
     public ServerRole getRole() throws ArangoDBException {
         return executor.execute(getRoleRequest(), getRoleResponseDeserializer());
+    }
+
+    @Override
+    public String getServerId() throws ArangoDBException {
+        return executor.execute(getServerIdRequest(), getServerIdResponseDeserializer());
     }
 
     @Override