+++ src/main/java/com/arangodb/internal/net/RandomHostHandler.java
@@ -30,27 +30,27 @@ public class RandomHostHandler implements HostHandler {
 
     private final HostResolver resolver;
     private final HostHandler fallback;
-    private Host origin;
     private Host current;
+    private HostSet hosts;
 
     public RandomHostHandler(final HostResolver resolver, final HostHandler fallback) {
         super();
         this.resolver = resolver;
         this.fallback = fallback;
-        origin = current = getRandomHost(true, false);
+        current = getRandomHost(true, false);
     }
 
     @Override
     public Host get(final HostHandle hostHandle, AccessType accessType) {
         if (current == null) {
-            origin = current = getRandomHost(false, true);
+            current = getRandomHost(false, true);
         }
         return current;
     }
 
     @Override
     public void success() {
-        current = origin;
+        fallback.success();
     }
 
     @Override
@@ -60,10 +60,10 @@ public class RandomHostHandler implements HostHandler {
     }
 
     private Host getRandomHost(final boolean initial, final boolean closeConnections) {
-
-        final ArrayList<Host> hosts = new ArrayList<>(resolver.resolve(initial, closeConnections).getHostsList());
-        Collections.shuffle(hosts);
-        return hosts.get(0);
+        hosts = resolver.resolve(initial, closeConnections);
+        final ArrayList<Host> hostList = new ArrayList<>(hosts.getHostsList());
+        Collections.shuffle(hostList);
+        return hostList.get(0);
     }
 
     @Override
@@ -77,7 +77,6 @@ public class RandomHostHandler implements HostHandler {
 
     @Override
     public void close() {
-        final HostSet hosts = resolver.resolve(false, false);
         hosts.close();
     }