+++ src/main/java/com/arangodb/internal/velocystream/VstCommunication.java
@@ -83,6 +83,7 @@ public abstract class VstCommunication<R, C extends VstConnection> implements Cl
             }
             final C connection = (C) host.connection();
             if (connection.isOpen()) {
+                hostHandler.success();
                 return connection;
             } else {
                 try {
@@ -92,6 +93,12 @@ public abstract class VstCommunication<R, C extends VstConnection> implements Cl
                         tryAuthenticate(connection);
                     }
                     hostHandler.confirm();
+                    if (!connection.isOpen()) {
+                        // see https://github.com/arangodb/arangodb-java-driver/issues/384
+                        hostHandler.fail();
+                        host = hostHandler.get(hostHandle, accessType);
+                        continue;
+                    }
                     return connection;
                 } catch (final IOException e) {
                     hostHandler.fail();